+++ packages/terra-popup/CHANGELOG.md
@@ -3,6 +3,10 @@ Changelog
 
 Unreleased
 ----------
+### Breaking Changes
+* Removed `releaseFocus` prop
+* Removed `requestFocus` prop
+* Updated focus-trap-react dependency to ^6.0.0
 
 5.12.0 - (May 7, 2019)
 ------------------